+++ src/App.css
@@ -93,47 +93,58 @@
   background-color: #6c5b7b;
   color: #f8b195;
 }
-.game-wrapper {
-  height: 800px;
-  width: 1000px;
-  margin: 3% auto;
-  position: relative;
-}
-.field {
+
+.ui-field {
   background-color: #6c5b7b;
-  height: 800px;
-  width: 800px;
+  height: 80vh;
+  width: 80vh;
   margin: auto;
   position: relative;
 }
+.game-field {
+  height: 95%;
+  width: 95%;
+  position: relative;
+  left: 2.5%;
+  top: 2.5%;
+}
 .dotted-line {
-  height: 800px;
-  width: 400px;
+  height: 100%;
+  width: 50%;
   border-right: 3px dashed #f8b195;
-  position: fixed;
+  position: absolute;
 }
 .ball {
   width: 5%;
   height: 5%;
   position: absolute;
-  background-color: #f8b195;
-  border-radius: 50%;
-}
-.right-card {
-  height: 30%;
-  width: 10%;
   background-color: #291e33;
-  position: absolute;
-  left: 0px;
+  border-radius: 50%;
+  transform: translate(-50%, -50%);
 }
 .left-card {
-  height: 30%;
+  height: 20vh;
   width: 10%;
   background-color: #291e33;
   position: absolute;
-  left: 900px;
+  left: -10%;
+  transform: translate(0, -50%);
+}
+.right-card {
+  height: 20vh;
+  width: 10%;
+  background-color: #291e33;
+  position: absolute;
+  left: 100%;
+  transform: translate(0, -50%);
 }
 .player-card-image {
   width: 90%;
   height: 100%;
+}
+.cards-wrapper {
+  height: 60vh;
+  width: 80vh;
+  position: absolute;
+  top: 10vh;
 }
